+++ requirements.txt
@@ -38,9 +38,6 @@ Mako==1.1.0
 Markdown==3.1.1
 MarkupSafe==1.1.1
 matplotlib==3.1.1
-mkl-fft==1.0.15
-mkl-random==1.1.0
-mkl-service==2.3.0
 mlflow==1.4.0
 networkx==2.4
 numpy==1.17.3
