+++ sapsan/config.py
@@ -1,3 +1,5 @@
+from typing import List, Optional
+
 from sapsan.Sapsan import Sapsan
 
 pars = {'name':'16','ttrain':[0]}
@@ -28,63 +30,4 @@ sps = Sapsan()
 model = sps.fit(pars)
 
 #sps.test(model, ttest=[1])
-
-
-class SapsanConfig(object):
-    def __init__(
-            self,
-            dim: int,
-            parameters: list, # TODO: rename to features or something like that :)
-            target: str,
-            filter_modes: int,
-            target_comp: int,
-            alpha: float,
-            gamma: float,
-            max_dim: int,
-            dataset: str,
-            axis: int,
-            from_3d: bool,
-            savepath: str,
-            dt: float,
-            data_type: str,
-            path: str,
-            method: str,
-            experiment_name: str
-    ):
-        """
-        :param dim: dimension of the training set
-        :param parameters: parameters (e.g. features) to train on
-        :param target: target parameter to train against and to predict
-        :param filter_modes: number of modes to filter down to
-        :param target_comp:
-        :param alpha: will use the default KRR alpha
-        :param gamma: will use the default KRR gamma
-        :param max_dim: initial dimension of the data file, from which training and testing sets will be extracted
-        :param dataset: which data set to use; options are 'iso' for HD turbulence, and 'mhd' for MHD turbulence
-        :param axis: dimensionality; either 2 or 3
-        :param from_3d: extracts 2D slice from 3D data; only relevant if axis=2
-        :param savepath: where to save the data
-        :param dt: numeral value for dt to calculate the actual time of each timestep
-        :param data_type:
-        :param path: path to the data
-        :param method: method to use for training; either 'cnn' or 'krr'
-        :param experiment_name: name of the experiment for mlflow tracking
-        """
-        self.dim = dim
-        self.parameters = parameters
-        self.target = target
-        self.filter_modes = filter_modes
-        self.target_comp = target_comp
-        self.alpha = alpha
-        self.gamma = gamma
-        self.max_dim = max_dim
-        self.dataset = dataset
-        self.axis = axis
-        self.from_3d = from_3d
-        self.savepath = savepath
-        self.dt = dt
-        self.data_type = data_type
-        self.path = path
-        self.method = method
-        self.experiment_name = experiment_name
