+++ sapsan/config/default.yaml
@@ -1,0 +1,33 @@
+general:
+  name: '16'
+  ttrain: [0]
+  dim: 16
+  parameters: [
+    'u_dim128_fm15',
+    'b_dim128_fm15',
+    'a_dim128_fm15',
+    'du0_dim128_fm15',
+    'du1_dim128_fm15',
+    'du2_dim128_fm15',
+    'db0_dim128_fm15',
+    'db1_dim128_fm15',
+    'db2_dim128_fm15',
+    'da0_dim128_fm15',
+    'da1_dim128_fm15',
+    'da2_dim128_fm15'
+  ]
+  target: 'tn_dim128_fm15'
+  filter_modes: 10
+  target_comp: 2
+  alpha: 0.001
+  gamma: 1.778
+  max_dim: 128
+  dataset: mhd
+  axis: 3
+  from_3d: true
+  savepath: 'Figures/'
+  dt: 2.5e-3
+  data_type: 'h5'
+  path: '/raid1/JHTDB/mhd/max_1024/mhd128_t@.4/'
+  method: 'cnn'
+  experiment_name: 'trial'
